wip - refactoring ml part
